requirements.txt
gradio>=3.20.6
python-dotenv>=1.0.0
